Refactor pretty much everything
