+++ Sources/TMDB/Configuration/ConfigurationClient.swift
@@ -1,12 +1,28 @@
+//
+//  ConfigurationClient.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
 import Dependencies
 import DependenciesMacros
-import Foundation
 
 @DependencyClient public struct ConfigurationClient: Sendable {
   var details:
     @Sendable (
-      _ accessToken: String
+      _ accessToken: AccessToken
     ) async throws -> ConfigurationDetails
+
+  var languages:
+    @Sendable (
+      _ accessToken: AccessToken
+    ) async throws -> [Language]
+
+  var primaryTranslations:
+    @Sendable (
+      _ accessToken: AccessToken
+    ) async throws -> [String]
 }
 
 extension ConfigurationClient: DependencyKey {
@@ -15,6 +31,24 @@ extension ConfigurationClient: DependencyKey {
       @Dependency(\.sharedClient) var sharedClient
       return try await sharedClient.fetch(
         relativePath: "configuration",
+        queryItems: nil,
+        locale: nil,
+        accessToken: accessToken
+      )
+    },
+    languages: { accessToken in
+      @Dependency(\.sharedClient) var sharedClient
+      return try await sharedClient.fetch(
+        relativePath: "configuration/languages",
+        queryItems: nil,
+        locale: nil,
+        accessToken: accessToken
+      )
+    },
+    primaryTranslations: { accessToken in
+      @Dependency(\.sharedClient) var sharedClient
+      return try await sharedClient.fetch(
+        relativePath: "configuration/primary_translations",
         queryItems: nil,
         locale: nil,
         accessToken: accessToken
