+++ Sources/TMDB/ImagePaths/PosterPath.swift
@@ -1,0 +1,24 @@
+//
+//  PosterPath.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+public struct PosterPath: Codable, Hashable, ImagePath, Sendable {
+  public static var keyPath: KeyPath<ImageConfiguration, [PosterSize]> {
+    \.posterSizes
+  }
+
+  public var rawValue: String
+
+  public init(rawValue: String) {
+    self.rawValue = rawValue
+  }
+}
+
+extension PosterPath: ExpressibleByStringLiteral {
+  public init(stringLiteral value: String) {
+    self.rawValue = value
+  }
+}
