+++ Sources/TMDB/ImageSizes/PosterSize.swift
@@ -1,0 +1,20 @@
+//
+//  PosterSize.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+public struct PosterSize: Codable, Hashable, RawRepresentable, Sendable {
+  public var rawValue: String
+
+  public init(rawValue: String) {
+    self.rawValue = rawValue
+  }
+}
+
+extension PosterSize: ExpressibleByStringLiteral {
+  public init(stringLiteral value: String) {
+    self.rawValue = value
+  }
+}
