+++ Sources/TMDB/ImageSizes/ProfileSize.swift
@@ -1,0 +1,20 @@
+//
+//  ProfileSize.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+public struct ProfileSize: Codable, Hashable, RawRepresentable, Sendable {
+  public var rawValue: String
+
+  public init(rawValue: String) {
+    self.rawValue = rawValue
+  }
+}
+
+extension ProfileSize: ExpressibleByStringLiteral {
+  public init(stringLiteral value: String) {
+    self.rawValue = value
+  }
+}
