+++ Sources/TMDB/Movies/MovieAppendOptions.swift
@@ -1,0 +1,11 @@
+//
+//  MovieAppendOptions.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+public enum MovieAppendOptions: String, CaseIterable, Codable, Hashable, Sendable {
+  case keywords
+  case similar
+}
