--- Sources/TMDB/Movies/MovieDetailsAppendingOptions.swift
@@ -1,4 +1,0 @@
-public enum MovieDetailsAppendingOptions: String, CaseIterable, Codable, Hashable, Sendable {
-  case keywords
-  case similar
-}
