+++ Sources/TMDB/Shared/Country.swift
@@ -1,9 +1,20 @@
-public struct Country: Codable, Hashable, Sendable {
-  public var iso31661 = ""
-  public var name = ""
+//
+//  Country.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
 
-  public init() {
-    // Makes initializer public
+public struct Country: Codable, Hashable, Sendable {
+  public var iso31661: String
+  public var name: String
+
+  public init(
+    iso31661: String,
+    name: String
+  ) {
+    self.iso31661 = iso31661
+    self.name = name
   }
 
   private enum CodingKeys: String, CodingKey {
