+++ Sources/TMDB/Shared/Language.swift
@@ -1,10 +1,23 @@
-public struct Language: Codable, Hashable, Sendable {
-  public var englishName = ""
-  public var iso6391 = ""
-  public var name = ""
+//
+//  Language.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
 
-  public init() {
-    // Makes initializer public
+public struct Language: Codable, Hashable, Sendable {
+  public var englishName: String
+  public var iso6391: String
+  public var name: String
+
+  public init(
+    englishName: String,
+    iso6391: String,
+    name: String
+  ) {
+    self.englishName = englishName
+    self.iso6391 = iso6391
+    self.name = name
   }
 
   public enum CodingKeys: String, CodingKey {
