--- Sources/TMDB/Shared/TMDBClientError.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public enum TMDBClientError: Error {
-  case invalidBaseURLString(String)
-  case invalidURLComponents(URLComponents)
-}
