+++ Sources/TMDB/Shared/TimeWindow.swift
@@ -1,0 +1,11 @@
+//
+//  TimeWindow.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+public enum TimeWindow: String, Codable, Hashable, Sendable {
+  case day
+  case week
+}
