--- Sources/TMDB/Shared/TrendingTimeWindow.swift
@@ -1,4 +1,0 @@
-public enum TrendingTimeWindow: String, Codable, Hashable, Sendable {
-  case day
-  case week
-}
