+++ Sources/TMDB/TVShows/TVShowAppendOptions.swift
@@ -1,0 +1,11 @@
+//
+//  TVShowAppendOptions.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
+public enum TVShowAppendOptions: String, CaseIterable, Codable, Hashable, Sendable {
+  case keywords
+  case similar
+}
