+++ Sources/TMDB/TVShows/TVShowClient.swift
@@ -1,27 +1,34 @@
+//
+//  TVShowClient.swift
+//  TMDB
+//
+//  Created by Igor Camilo on 13.08.25.
+//
+
 import Dependencies
 import DependenciesMacros
 import Foundation
 
 @DependencyClient struct TVShowClient {
-  var tvShowDetails:
+  var details:
     @Sendable (
       _ id: TVShow.ID,
-      _ appending: [TVShowDetailsAppendingOptions],
+      _ appending: [TVShowAppendOptions],
       _ locale: Locale?,
-      _ accessToken: String
+      _ accessToken: AccessToken
     ) async throws -> TVShowDetails
 
   var tvShows:
     @Sendable (
       _ list: TVShowList,
       _ locale: Locale?,
-      _ accessToken: String
+      _ accessToken: AccessToken
     ) async throws -> Page<TVShow>
 }
 
 extension TVShowClient: DependencyKey {
   static let liveValue = TVShowClient(
-    tvShowDetails: { id, appending, locale, accessToken in
+    details: { id, appending, locale, accessToken in
       var queryItems: [URLQueryItem] = []
       if !appending.isEmpty {
         let value = appending.map(\.rawValue).joined(separator: ",")
