--- Sources/TMDB/TVShows/TVShowDetailsAppendingOptions.swift
@@ -1,4 +1,0 @@
-public enum TVShowDetailsAppendingOptions: String, CaseIterable, Codable, Hashable, Sendable {
-  case keywords
-  case similar
-}
